requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7
boto3==1.28.85
python-dotenv==1.0.0
Pillow==10.0.1
PyPDF2==3.0.1
cloudinary==1.41.0
vercel-wsgi==0.8.2
psycopg2-binary==2.9.9
